refactor: Replace glob tool with grep tool in advanced example
- Replace GlobTool with GrepTool for content search using regex patterns
- Use bash grep commands instead of file pattern matching
- Support file filtering with include parameter (e.g., '*.py', '*.{ts,tsx}')
- Return matching lines with file paths and line numbers
- Demonstrate proper bash executor reuse pattern
- Update example conversation to search for 'class' keyword in Python files
- Rename file from advanced_tools_with_glob.py to advanced_tools_with_grep.py
